Attach an input to RichFuture.Action, and keep it in executeSequentially results
Define ActionTry (with ActionSuccess and ActionFailure) as action result, holding action input and actual result (with associated implicit conversion to Try/Success/Failure).
In case of action failure, allows to determine more easily which input is concerned, instead of having to keep separately the list of inputs and cross-checking with results.