RTFM and use setDisable instead of setDisabled
setDisabled should only be used to enforce the internal rules of JavaFX: e.g. a parent Node which we disable (with 'setDisable') shall disable all its children (with 'setDisabled').
This fixes the error preventing to use a trait instead of an abstract class.
Moreover 'setDisable' was already called when displaying a cell without data (e.g. separator), which would trigger unexpected behaviour: cells for which a separator is displayed then an item have 'disable' to true (preventing to change the checkbox state) and 'disabled' to false (which makes it rendered as if the checkbox state can be changed).