Add 'portable settings' based entirely on Config
Basically uses reference.conf as reference (fallback) and a given file
path (e.g. application.conf) for the application conf file.
Entries can be built based on those with get/set helpers.
Changing an entry value updates the underlying Config and application
conf file. File is backuped on the first change; the backup is used as
default if the application conf file cannot be parsed upon loading.

This can be used as a portable application with its configuration saved
in the application path, which would be easier to manually access than
the persistence preferences which are stored in registry under Windows,
or Java-proprietary path under Linux.