Make Alert Dialog headerText really optional
In simple alerts, the actual minimum is to set the content text.
Setting header text adds a visual separation (between the header and
content + buttons), which usually is not needed/wanted.
Header text was already Option, but have a default value of None so that
callers don't have to explicitly give None.