Process command arguments asynchronously even for the first unique instance
From the caller point of view, this makes all command arguments (from this
instance and from other instances) processed in a similar manner.

This is also useful for JavaFX applications: we usually need to wait for
the GUI to be 'ready' (e.g. main controller built) before processing any
command arguments, and the GUI can only start when 'launch'ing the JavaFX
Application which is a function that only returns upon exiting the
application (that is, the unique instance must 'start' before 'launch'ing
JavaFX).
In this case, caller gives a Future (derived from a Promise) that would be
completed once the GUI is ready (usually inside the JavaFX Application
'start' function).