Always keep portable settings reference configuration as fallback
Explicitly separate the standalone configuration (which we load, modify
and save) from the reference one. This also reduce the number of entries
to process when saving.
We never actually 'remove' any entry: at best we drop it from the
standalone configuration to fallback to the reference value. Make sure
this is the case (in both portable settings and config entries that can
be built out of it), and rename the 'remove' function to 'reset' in config
entry class to show the real intent.