Reset config entry if value is the same as reference value
Entry was already resetted upon setting an (explicitly defined) default
value. Do the same when value is the same as in the reference configuration.
That way underlying portable settings are automatically 'cleaned':
only changed values (or empty objects) should remain.

This also helps when configuration value representation is different
from actual logical value, as is the case for durations: represented as
string "60 seconds" is different from "1 minute", while the duration
itself is the same.
There is no more need for the dummy portable settings (built on the
reference configuration only), as default (reference) values are now
directly accessible through each config entry object.