Handle reference.conf overriding more properly
There is no guarantee of the order in which reference.conf files are merged.
So if there is a need to override values belonging to other libraries (say
akka) it has to be manually handled by loading and ordering each layer:
system properties overrides application.conf which overrides our
reference-overrides.conf which overrides reference.conf.