Prefer extending Enumeration over redefining it
This may involve a bit more/different code when defining an Enumeration,
but is enough for what it is used for.
This simplifies a bit the core code, and we can also drop associated
classes, functions and tests.

LogLevel reworked to be a full-fledged Enumeration with necessary
extensions.
RichEnumeration.byName also reworked to take into account extensions.