Command: automatically make stdin interruptible
Reading from stdin is blocking and cannot be interrupted (unless exiting
the JVM). When a spawn process is done, if there was no need for input
data, the input filtering code still remain blocked on stdin until data
is finally available or we exit the JVM.

Both cases are problematic:
If we read data after te command is done, it should be data meant for
another command.
Otherwise the resource remain present until the JVM stops.